Language.PCPL.Solver

Description

A simple PCP solver. There is a lot of potential for optimization here.

Documentation

type PCP = [(Int, Domino)]

PCP instance

data Node

Node in the search tree

Constructors

Node [Int] Configuration

data Configuration

The unmatched portion of an intermediate PCP state. The paper Tackling Post's Correspondence Problem calls this a configuration.

Constructors

Top [Symbol]
Portion of the top string that extends past the bottom one

Bottom [Symbol]
Portion of the bottom string that extends past the top one

search :: PCP -> [Node] -> [Int]

Find a solution using BFS.

updateConf :: Configuration -> Domino -> Maybe Configuration

Try to update the Configuration with the given Domino.